refactor: move team management logic to separate file
Extract all team-related code (TeamMemberVM, polling, avatar loading, presence announcements) from MainWindow.xaml.cs into a dedicated partial class file MainWindow.Team.Core.cs. This improves maintainability and separation of concerns by isolating the team feature logic.
