+++ RustPlusDesktop/MainWindow.xaml.cs
@@ -448,122 +448,6 @@ public partial class MainWindow : Window
             AppendLog("dyn2 error: " + ex.Message);
         }
     }
-    private System.Windows.Threading.DispatcherTimer? _teamTimer;
-    public ObservableCollection<TeamMemberVM> TeamMembers { get; } = new();
-
-    private int _teamBusy = 0;
-    //------------- TEAM UI --------------------------
-    private bool _iAmLeader = false;
-
-    // optional: Avatar-Cache, damit wir nicht dauernd laden
-    private readonly Dictionary<ulong, ImageSource> _avatarCache = new();
-    // Zugriff auf die echte Client-Klasse
-    private RustPlusClientReal? _real => _rust as RustPlusClientReal;
-
-    // Liste, die das Team-Tab anzeigt
-
-    public sealed class TeamMemberVM : INotifyPropertyChanged
-    {
-        public event PropertyChangedEventHandler? PropertyChanged;
-        private void OnChanged(string name) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-
-        public int MissingCount { get; set; }
-
-        // Schlüssel / Stammdaten
-        public ulong SteamId { get; init; }
-
-        private string _name = "(player)";
-        public string Name
-        {
-            get => _name;
-            set { if (_name == value) return; _name = value; OnChanged(nameof(Name)); }
-        }
-
-        private bool _isLeader;
-        public bool IsLeader
-        {
-            get => _isLeader;
-            set { if (_isLeader == value) return; _isLeader = value; OnChanged(nameof(IsLeader)); }
-        }
-
-        private bool _isOnline;
-        public bool IsOnline
-        {
-            get => _isOnline;
-            set
-            {
-                if (_isOnline == value) return;
-                _isOnline = value;
-                OnChanged(nameof(IsOnline));
-                OnChanged(nameof(IsOnlineAndAlive)); // abgeleitet
-            }
-        }
-
-        private bool _isDead;
-        public bool IsDead
-        {
-            get => _isDead;
-            set
-            {
-                if (_isDead == value) return;
-                _isDead = value;
-                OnChanged(nameof(IsDead));
-                OnChanged(nameof(IsOnlineAndAlive)); // abgeleitet
-            }
-        }
-
-        // Abgeleitet: nur dann grün, wenn online und nicht tot
-        public bool IsOnlineAndAlive => IsOnline && !IsDead;
-
-        public double? X { get; set; }
-        public double? Y { get; set; }
-
-        private ImageSource? _avatar;
-        public ImageSource? Avatar
-        {
-            get => _avatar;
-            set { if (_avatar == value) return; _avatar = value; OnChanged(nameof(Avatar)); }
-        }
-    }
-
-    // Letzte bekannten Player-Positionen (aus Dyn-Markern)
-    private readonly Dictionary<ulong, (double x, double y, string name)> _lastPlayersBySid = new();
-    // Merker für letzte Präsenz zum Melden von Änderungen
-    private readonly Dictionary<ulong, (bool online, bool dead)> _lastPresence = new();
-
-    // eigenes SteamId64 robust
-    private ulong _mySteamId => (ulong.TryParse(_vm?.SteamId64, out var v) ? v : 0UL);
-
-
-    private void StartTeamPolling()
-    {
-        if (_teamTimer != null) return; // schon aktiv
-        _teamTimer = new System.Windows.Threading.DispatcherTimer
-        {
-            Interval = TimeSpan.FromSeconds(5)
-        };
-        _teamTimer.Tick += TeamTimer_Tick;
-        _teamTimer.Start();
-    }
-
-    private void StopTeamPolling()
-    {
-        var t = _teamTimer;
-        if (t == null) return;
-        t.Tick -= TeamTimer_Tick;
-        t.Stop();
-        _teamTimer = null;
-    }
-
-    private int _teamPollBusy = 0;
-    private async void TeamTimer_Tick(object? sender, EventArgs e)
-    {
-        if (System.Threading.Interlocked.Exchange(ref _teamPollBusy, 1) == 1) return;
-        try { await LoadTeamAsync(); }
-        finally { System.Threading.Interlocked.Exchange(ref _teamPollBusy, 0); }
-        CenterMiniMapOnPlayer();
-    }
-
     public class BoolToVisibilityConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
@@ -576,354 +460,6 @@ public partial class MainWindow : Window
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => Binding.DoNothing;
     }
 
-    private static object? P(object? o, string name) =>
-    o?.GetType().GetProperty(name,
-        BindingFlags.Instance | BindingFlags.Public | BindingFlags.IgnoreCase)?.GetValue(o);
-
-    private static T Get<T>(object? o, T def, params string[] names)
-    {
-        foreach (var n in names)
-        {
-            var v = P(o, n);
-            if (v is null) continue;
-            try
-            {
-                if (typeof(T) == typeof(string)) return (T)(object)(v.ToString() ?? "");
-                return (T)Convert.ChangeType(v, typeof(T), CultureInfo.InvariantCulture);
-            }
-            catch { }
-        }
-        return def;
-    }
-
-    // RAW Fallback for Team online info and Leader Info
-
-    private async Task EnsureAvatarAsync(TeamMemberVM vm)
-    {
-        if (!_avatarLoading.Add(vm.SteamId)) return;
-        try
-        {
-            // 1) Laden (nutzt DEINE vorhandene Methode)
-            await LoadAvatarAsync(vm).ConfigureAwait(false);
-
-            // 2) Wenn erfolgreich: Overlay sofort auf Avatar umschalten
-            if (vm.Avatar != null)
-            {
-                await Dispatcher.InvokeAsync(() =>
-                {
-                    // Alle Marker dieses Spielers aktualisieren
-                    foreach (var kv in _dynEls.ToList())
-                    {
-                        if (kv.Value is FrameworkElement fe &&
-                            fe.Tag is PlayerMarkerTag t &&
-                            t.SteamId == vm.SteamId)
-                        {
-                            var el = fe; // ref-Arg
-                            UpdatePlayerMarker(ref el, kv.Key, vm.SteamId, vm.Name, vm.IsOnline, vm.IsDead);
-                            // Skalierung bleibt konsistent:
-                            ApplyCurrentOverlayScale(el);
-                        }
-                    }
-                });
-                // erfolgreicher Load -> Retry-Fenster löschen
-                _avatarNextTry.Remove(vm.SteamId);
-            }
-            else
-            {
-                // kein Bild bekommen -> später nochmal probieren
-                _avatarNextTry[vm.SteamId] = DateTime.UtcNow + AvatarRetryInterval;
-            }
-        }
-        catch
-        {
-            // Fehler -> später nochmal probieren
-            _avatarNextTry[vm.SteamId] = DateTime.UtcNow + AvatarRetryInterval;
-        }
-        finally
-        {
-            _avatarLoading.Remove(vm.SteamId);
-        }
-    }
-
-
-    private async Task LoadTeamAsync()
-    {
-        if (_real is null) return;
-
-        try
-        {
-            var team = await _real.GetTeamInfoAsync();
-            if (team is null) return;
-
-            var leaderId = team.LeaderSteamId;
-            var seen = new HashSet<ulong>();
-            foreach (var m in TeamMembers) m.MissingCount++;
-
-            foreach (var m in team.Members)
-            {
-                var sid = m.SteamId;
-                if (sid == 0) continue;
-                seen.Add(sid);
-
-                var vm = TeamMembers.FirstOrDefault(t => t.SteamId == sid);
-                if (vm == null)
-                {
-                    vm = new TeamMemberVM { SteamId = sid };
-                    TeamMembers.Add(vm);
-                    _ = LoadAvatarAsync(vm); // nutzt deine vorhandene Avatar-Funktion
-                    if (vm.Avatar == null && CanTryAvatar(sid))
-                    {
-                        _ = EnsureAvatarAsync(vm); // siehe unten
-                    }
-                }
-                vm.MissingCount = 0;
-
-                // Vorheriger Zustand für Announcements
-                var hadPrev = _lastPresence.TryGetValue(sid, out var prev);
-
-                vm.Name = string.IsNullOrWhiteSpace(m.Name) ? "(player)" : m.Name!;
-                vm.IsLeader = (leaderId != 0 && sid == leaderId);
-                vm.IsOnline = m.Online;
-                vm.IsDead = m.Dead;
-                vm.X = m.X; vm.Y = m.Y;
-
-                var now = (m.Online, m.Dead);
-                _lastPresence[sid] = now;
-
-                if (hadPrev && prev != now)
-                    _ = AnnouncePresenceChangeAsync(vm, prev, now);
-
-            }
-
-            // Entferne nicht mehr vorhandene (nach 10 Sek Puffer)
-            for (int i = TeamMembers.Count - 1; i >= 0; i--)
-                if (TeamMembers[i].MissingCount > 2)
-                    TeamMembers.RemoveAt(i);
-
-            // Log hilft bei der Rechteprüfung LEADER etc.
-            //  var iAmLeader = TeamMembers.Any(t => t.SteamId == _mySteamId && t.IsLeader);
-            // AppendLog($"[team] leader={leaderId} me={_mySteamId} -> iAmLeader={iAmLeader} members={TeamMembers.Count}");
-        }
-        catch (Exception ex)
-        {
-            AppendLog("[team] " + ex.Message);
-        }
-
-        if (_overlayToolsVisible)
-        {
-            // UI muss auf UI-Thread aktualisiert werden
-            await Dispatcher.InvokeAsync(() =>
-            {
-                RebuildOverlayTeamBar();
-            });
-        }
-
-    }
-
-    private async Task AnnouncePresenceChangeAsync(TeamMemberVM vm, (bool online, bool dead) prev, (bool online, bool dead) now)
-    {
-        try
-        {
-            // Online/Offline (wenn dieser Part auch vom Button abhängen soll, ebenfalls mit _announceSpawns verknüpfen)
-            if (prev.online != now.online  && _announceSpawns )
-            {
-                bool isSelf = (vm.SteamId == _mySteamId);
-                bool shouldAnnounce = now.online ? TrackingService.AnnouncePlayerOnline : TrackingService.AnnouncePlayerOffline;
-
-                if (shouldAnnounce)
-                {
-                    var where = (vm.X.HasValue && vm.Y.HasValue) ? GetGridLabel(vm.X.Value, vm.Y.Value) : "unknown";
-                    var txt = now.online ? $"{vm.Name} came online @ {where}"
-                                         : $"{vm.Name} went offline";
-                    await SendTeamChatSafeAsync(txt);
-                }
-            }
-
-            // Death/Respawn
-            if (prev.dead != now.dead)
-            {
-                // Position robust bestimmen (immer versuchen, unabhängig vom Toggle)
-                double? px = vm.X, py = vm.Y;
-                if ((!px.HasValue || !py.HasValue) && TryResolvePosFromDynMarkers(vm.SteamId, out var dx, out var dy))
-                { px = dx; py = dy; }
-
-                // ---> Meldung nur senden, wenn Button aktiv
-                if (_announceSpawns)
-                {
-                    bool isSelf = (vm.SteamId == _mySteamId);
-                    bool shouldAnnounce = false;
-
-                    if (prev.dead != now.dead)
-                    {
-                        if (now.dead) shouldAnnounce = isSelf ? TrackingService.AnnouncePlayerDeathSelf : TrackingService.AnnouncePlayerDeathTeam;
-                        else shouldAnnounce = isSelf ? TrackingService.AnnouncePlayerRespawnSelf : TrackingService.AnnouncePlayerRespawnTeam;
-                    }
-
-                    if (shouldAnnounce)
-                    {
-                        var where = (px.HasValue && py.HasValue) ? GetGridLabel(px.Value, py.Value) : "unknown";
-                        var txt = now.dead ? $"{vm.Name} died @ {where}" : $"{vm.Name} respawned @ {where}";
-                        await SendTeamChatSafeAsync(txt);
-                    }
-                }
-
-                // Death-Pin weiter unabhängig vom Toggle behandeln
-                if (_showDeathMarkers && now.dead && px.HasValue && py.HasValue)
-                {
-                    PlaceOrMoveDeathPin(vm.SteamId, px.Value, py.Value, vm.Name);
-                }
-                else if (!now.dead)
-                {
-                    // RemoveDeathPins(vm.SteamId);
-                }
-            }
-        }
-        catch { }
-    }
-
-
-
-    // Avatar laden (ohne API-Key; via ?xml=1) + Cache
-    private static readonly HttpClient _http = new HttpClient(new HttpClientHandler
-    {
-        AutomaticDecompression = System.Net.DecompressionMethods.All
-    });
-
-
-    private static ImageSource? BytesToImage(byte[] bytes)
-    {
-        try
-        {
-            var bi = new BitmapImage();
-            using var ms = new MemoryStream(bytes);
-            bi.BeginInit();
-            bi.CacheOption = BitmapCacheOption.OnLoad;
-            bi.StreamSource = ms;
-            bi.EndInit();
-            bi.Freeze();
-            return bi;
-        }
-        catch { return null; }
-    }
-
-    private static async Task<ImageSource?> FetchSteamAvatarAsync(ulong steamId)
-    {
-        if (steamId == 0) return null;
-        try
-        {
-            using var http = new HttpClient();
-            // 1) Avatar-URL aus dem XML-Profil ziehen (braucht keinen API-Key)
-            var xml = await http.GetStringAsync($"https://steamcommunity.com/profiles/{steamId}?xml=1");
-            string url = "";
-            var mFull = Regex.Match(xml, @"<avatarFull><!\[CDATA\[(.*?)\]\]></avatarFull>", RegexOptions.IgnoreCase);
-            var mMedium = Regex.Match(xml, @"<avatarMedium><!\[CDATA\[(.*?)\]\]></avatarMedium>", RegexOptions.IgnoreCase);
-            if (mFull.Success) url = mFull.Groups[1].Value;
-            else if (mMedium.Success) url = mMedium.Groups[1].Value;
-            if (string.IsNullOrWhiteSpace(url)) return null;
-
-            // 2) Bild laden
-            var bytes = await http.GetByteArrayAsync(url);
-            return BytesToImage(bytes);
-        }
-        catch { return null; }
-    }
-
-    private async Task LoadAvatarAsync(TeamMemberVM vm)
-    {
-        try
-        {
-            if (vm.SteamId == 0 || vm.Avatar != null) return;
-
-            if (_avatarCache.TryGetValue(vm.SteamId, out var cached) && cached != null)
-            { vm.Avatar = cached; return; }
-
-            var img = await FetchSteamAvatarAsync(vm.SteamId);
-            if (img != null)
-            {
-                _avatarCache[vm.SteamId] = img;
-                vm.Avatar = img;
-            }
-        }
-        catch (Exception ex)
-        {
-            AppendLog($"[avatar] {vm.SteamId}: {ex.Message}");
-        }
-    }
-
-    // Linksklick: auf Karte zentrieren
-    private void TeamItem_MouseLeftButtonUp(object sender, MouseButtonEventArgs e)
-    {
-        if ((sender as FrameworkElement)?.DataContext is TeamMemberVM vm)
-            CenterOnMember(vm);
-    }
-
-    // Kontextmenü – Zentrieren
-    private void Team_Center_Click(object sender, RoutedEventArgs e)
-    {
-        if ((sender as FrameworkElement)?.DataContext is TeamMemberVM vm)
-            CenterOnMember(vm);
-    }
-
-    private void Team_OpenProfile_Click(object sender, RoutedEventArgs e)
-    {
-        var vm = VMFromSender(sender); if (vm == null) return;
-        try
-        {
-            var url = $"https://steamcommunity.com/profiles/{vm.SteamId}";
-            System.Diagnostics.Process.Start(new System.Diagnostics.ProcessStartInfo(url) { UseShellExecute = true });
-        }
-        catch { /* ignore */ }
-    }
-
-    // Kontextmenü – Promote
-    private bool IAmLeaderNow() => TeamMembers.Any(t => t.SteamId == _mySteamId && t.IsLeader);
-    private TeamMemberVM? VMFromSender(object sender)
-        => (sender as FrameworkElement)?.DataContext as TeamMemberVM ?? TeamList?.SelectedItem as TeamMemberVM;
-
-    private async void Team_Promote_Click(object sender, RoutedEventArgs e)
-    {
-        var vm = VMFromSender(sender); if (vm == null) return;
-        if (!IAmLeaderNow()) { AppendLog("Only Leader can promote."); return; }
-        if (vm.SteamId == _mySteamId) return;
-        try { await (_real as RustPlusClientReal)?.PromoteToLeaderAsync(vm.SteamId); }
-        catch (Exception ex) { AppendLog("[team] promote error: " + ex.Message); }
-    }
-
-
-
-    private void CenterOnMember(TeamMemberVM vm)
-    {
-        if (vm.X.HasValue && vm.Y.HasValue)
-        {
-            CenterMapOnWorld(vm.X.Value, vm.Y.Value);
-            return;
-        }
-        if (TryResolvePosFromDynMarkers(vm.SteamId, out var x, out var y))
-        {
-            CenterMapOnWorld(x, y);
-            return;
-        }
-        MessageBox.Show("Keine Position verfügbar (offline oder nicht gespawnt).");
-    }
-
-    private bool TryResolvePosFromDynMarkers(ulong sid, out double x, out double y)
-    {
-        if (_lastPlayersBySid.TryGetValue(sid, out var pos))
-        {
-            x = pos.x; y = pos.y; return true;
-        }
-        x = y = 0; return false;
-    }
-
-    // Kontextmenü – Kick
-    private async void Team_Kick_Click(object sender, RoutedEventArgs e)
-    {
-        var vm = VMFromSender(sender); if (vm == null) return;
-        if (!IAmLeaderNow()) { AppendLog("Only Leader can kick."); return; }
-        if (vm.SteamId == _mySteamId) return;
-        try { await (_real as RustPlusClientReal)?.KickTeamMemberAsync(vm.SteamId); }
-        catch (Exception ex) { AppendLog("[team] kick error: " + ex.Message); }
-    }
     public static class AppInfo
     {
         public static string VersionRaw =>
@@ -950,11 +486,6 @@ public partial class MainWindow : Window
 
     private BitmapSource? _mapBaseBmp; // Original-Map ohne Marker
     private readonly List<(double uPx, double vPx, string? label)> _staticMarkers = new();
-    // SteamId -> Name Cache
-    private readonly Dictionary<ulong, string> _steamNames = new();
-
-    // throttle: nicht bei jedem Tick refreshe n
-    private DateTime _lastTeamRefresh = DateTime.MinValue;
     protected override void OnClosing(System.ComponentModel.CancelEventArgs e)
     {
         if (TrackingService.CloseToTrayEnabled)
@@ -1680,23 +1211,6 @@ public partial class MainWindow : Window
         { }
     }
 
-    private string ResolvePlayerName(RustPlusClientReal.DynMarker m)
-    {
-        // 1) schon befüllt?
-        if (!string.IsNullOrWhiteSpace(m.Name)) return m.Name;
-        if (!string.IsNullOrWhiteSpace(m.Label)) return m.Label;
-
-        // 2) Cache nach steamId
-        if (m.SteamId != 0 && _steamNames.TryGetValue(m.SteamId, out var n) && !string.IsNullOrWhiteSpace(n))
-            return n;
-
-        // 3) lazy refresh anstoßen (nicht blockierend)
-        if (DateTime.UtcNow - _lastTeamRefresh > TimeSpan.FromSeconds(5))
-            _ = RefreshTeamNamesAsync();
-
-        return "(player)";
-    }
-
     // --- Chat Persistence & Switching ---
 
     private ServerProfile? _lastChatProfile;
@@ -1824,30 +1338,6 @@ public partial class MainWindow : Window
         var author = (m.Author ?? "").Trim().ToLowerInvariant();
         var text = (m.Text ?? "").Trim().ToLowerInvariant();
         return $"{author}|{text}";
-    }
-
-    private async Task RefreshTeamNamesAsync()
-    {
-        _lastTeamRefresh = DateTime.UtcNow;
-
-        if (_rust is not RustPlusClientReal real) return;
-
-        try
-        {
-            var team = await real.GetTeamInfoAsync(); // <— deine bestehende Methode
-            if (team?.Members != null)
-            {
-                foreach (var m in team.Members)
-                {
-                    if (m.SteamId != 0 && !string.IsNullOrWhiteSpace(m.Name))
-                        _steamNames[m.SteamId] = m.Name!;
-                }
-            }
-        }
-        catch
-        {
-            // still ok
-        }
     }
 
     private sealed class ItemInfo
